Added VS solution file and project files Fixed index out of bound bug when game hits last tetris piece.
