+++ Console-Tetris.sln
@@ -1,0 +1,31 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 16
+VisualStudioVersion = 16.0.29123.88
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "Console-Tetris", "Console-Tetris.vcxproj", "{577F129B-8764-4266-8DEF-3A5C351BA40C}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|x64 = Debug|x64
+		Debug|x86 = Debug|x86
+		Release|x64 = Release|x64
+		Release|x86 = Release|x86
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{577F129B-8764-4266-8DEF-3A5C351BA40C}.Debug|x64.ActiveCfg = Debug|x64
+		{577F129B-8764-4266-8DEF-3A5C351BA40C}.Debug|x64.Build.0 = Debug|x64
+		{577F129B-8764-4266-8DEF-3A5C351BA40C}.Debug|x86.ActiveCfg = Debug|Win32
+		{577F129B-8764-4266-8DEF-3A5C351BA40C}.Debug|x86.Build.0 = Debug|Win32
+		{577F129B-8764-4266-8DEF-3A5C351BA40C}.Release|x64.ActiveCfg = Release|x64
+		{577F129B-8764-4266-8DEF-3A5C351BA40C}.Release|x64.Build.0 = Release|x64
+		{577F129B-8764-4266-8DEF-3A5C351BA40C}.Release|x86.ActiveCfg = Release|Win32
+		{577F129B-8764-4266-8DEF-3A5C351BA40C}.Release|x86.Build.0 = Release|Win32
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {E2374F5B-22BB-4EC3-B7DC-20F0E95466AC}
+	EndGlobalSection
+EndGlobal
